+++ metrics/mypy_high_water_mark
@@ -1,1 +1,1 @@
-92.3300
+92.3500
